Hide edit/delete links based on canEdit(), etc.
This is a quick-n'-dirty patch to make DOM hide "action" links such as edit and delete based on per-item permissions (i.e., addresses issue #75). While a more complete solution is needed (there are other points in the code that should be updated), it does only show the action links when those actions are allowed.